Reorganized the files
 Did it for improved clarity and structure.
