+++ transformer_1/orel/model/config.py
@@ -1,0 +1,6 @@
+MAX_TANSOR_LENGTH = 15
+BATCH_SIZE = 64
+TEST_SIZE = 0.10
+VALIDATION_SIZE = 0.15
+EPOCHS = 3
+DROPOUT = 0.1
